requirements.txt
asgiref==3.2.3
boto3==1.14.12
botocore==1.17.12
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.7
django-analytical==2.5.0
django-cleanup==5.0.0
django-crispy==1.8.1
django-import-export==2.0.1
django-rest-knox==4.1.0
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
docutils==0.15.2
dynaconf==2.2.3
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.0.1
hvac==0.10.4
idna==2.8
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
loguru==0.5.1
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
pycparser==2.20
pycryptodome==3.9.6
PyPDF2==1.26.0
python-box==3.4.6
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
PyYAML==5.3
reportlab==3.5.34
requests==2.22.0
s3transfer==0.3.3
shortuuid==1.0.1
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
toml==0.10.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.1.0
win32-setctime==1.0.1
xhtml2pdf==0.2.4
mysqlclient==2.0.3